requirements_embeddings.txt
sentence-transformers
torch
supabase
pandas
numpy
tqdm
python-dotenv
